requirements.txt
streamlit>=1.30.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
gdown>=5.0.0
